Issue #2547 - fix RHEL6 nativeLib build in 14.1.1
Former-commit-id: a51231058aaf114bf8964ca067186919a9856a9a [formerly 7af6c4794f9523e6f0d632b0a9f34e8f5cbcda3f]
Former-commit-id: e3639da7031a4b349f266f02c2daa6dee9da617a

+++ nativeLib/rary.wfoapi.tmcp/src/tmcp.C
@@ -900,7 +900,7 @@ static void events_handler( Widget w, XtPointer client_data, XtPointer call_data
 {
     int rc;
     int	pc;  // pre-condition flag
-    int which = (int) client_data;
+    int which = (long) client_data;
     XmString label;
     XmString button_label;
     char 	command[512];
